[REFACTORING] Split EmailSetMethod and extract create/update/destroy
